--- src/main/java/EaseMood/com/api/ApiApplication.java
@@ -1,13 +1,0 @@
-package EaseMood.com.api;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApiApplication.class, args);
-	}
-
-}
