--- src/main/java/EaseMood/com/api/DTO/annotation/UniqueEmailValidator.java
@@ -1,17 +1,0 @@
-package EaseMood.com.api.DTO.annotation;
-
-import EaseMood.com.api.Repository.UserRepository;
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-import org.springframework.beans.factory.annotation.Autowired;
-
-public class UniqueEmailValidator implements ConstraintValidator<UniqueEmail, String> {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    @Override
-    public boolean isValid(String email, ConstraintValidatorContext context) {
-        return email != null && !userRepository.existsByEmail(email);
-    }
-}
