--- src/main/java/EaseMood/com/api/Repository/UserRepository.java
@@ -1,11 +1,0 @@
-package EaseMood.com.api.Repository;
-
-import EaseMood.com.api.Entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    boolean existsByEmail(String email);
-}
