--- src/main/java/EaseMood/com/api/Srvice/BeckDepressionInventoryService.java
@@ -1,10 +1,0 @@
-package EaseMood.com.api.Srvice;
-
-import EaseMood.com.api.Entity.BeckDepressionInventory;
-import EaseMood.com.api.Entity.User;
-import EaseMood.com.api.Repository.UserRepository;
-
-public interface BeckDepressionInventoryService {
-
-    int getResult(BeckDepressionInventory inventory, User user);
-}
