--- src/main/java/EaseMood/com/api/exception/DuplicateFieldException.java
@@ -1,8 +1,0 @@
-package EaseMood.com.api.exception;
-
-public class DuplicateFieldException extends RuntimeException {
-
-    public DuplicateFieldException(String message) {
-        super(message);
-    }
-}
