--- src/main/java/EaseMood/com/api/exception/ErrorResponse.java
@@ -1,6 +1,0 @@
-package EaseMood.com.api.exception;
-
-import java.util.List;
-public record ErrorResponse(int status, String message, List<String> errors) {
-
-}
