--- src/main/java/EaseMood/com/api/exception/NullEntityException.java
@@ -1,11 +1,0 @@
-package EaseMood.com.api.exception;
-
-public class NullEntityException extends RuntimeException {
-
-    public NullEntityException() {
-    }
-
-    public NullEntityException(String message) {
-        super(message);
-    }
-}
