+++ src/main/java/easy/mind/com/api/ApiApplication.java
@@ -1,0 +1,13 @@
+package easy.mind.com.api;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApiApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ApiApplication.class, args);
+	}
+
+}
