+++ src/main/java/easy/mind/com/api/DTO/annotation/UniqueEmailValidator.java
@@ -1,0 +1,17 @@
+package easy.mind.com.api.DTO.annotation;
+
+import easy.mind.com.api.Repository.UserRepository;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class UniqueEmailValidator implements ConstraintValidator<UniqueEmail, String> {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @Override
+    public boolean isValid(String email, ConstraintValidatorContext context) {
+        return email != null && !userRepository.existsByEmail(email);
+    }
+}
