+++ src/main/java/easy/mind/com/api/Entity/User.java
@@ -1,0 +1,32 @@
+package easy.mind.com.api.Entity;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+
+@Entity
+@Data
+@Builder
+@NoArgsConstructor
+@Table(name = "user_table")
+@AllArgsConstructor
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(name = "first_name", nullable = false)
+    private String firstName;
+
+
+    @Column(name = "last_name", nullable = false)
+    private String lastName;
+
+    @Column(name = "password", nullable = false)
+    private String password;
+
+    @Column(name = "email", nullable = false, unique = true)
+    private String email;
+
+}
