+++ src/main/java/easy/mind/com/api/Repository/UserRepository.java
@@ -1,0 +1,11 @@
+package easy.mind.com.api.Repository;
+
+import easy.mind.com.api.Entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+}
