+++ src/main/java/easy/mind/com/api/Srvice/BeckDepressionInventoryService.java
@@ -1,0 +1,9 @@
+package easy.mind.com.api.Srvice;
+
+import easy.mind.com.api.Entity.BeckDepressionInventory;
+import easy.mind.com.api.Entity.User;
+
+public interface BeckDepressionInventoryService {
+
+    int getResult(BeckDepressionInventory inventory, User user);
+}
