+++ src/main/java/easy/mind/com/api/exception/DuplicateFieldException.java
@@ -1,0 +1,8 @@
+package easy.mind.com.api.exception;
+
+public class DuplicateFieldException extends RuntimeException {
+
+    public DuplicateFieldException(String message) {
+        super(message);
+    }
+}
