+++ src/main/java/easy/mind/com/api/exception/ErrorResponse.java
@@ -1,0 +1,6 @@
+package easy.mind.com.api.exception;
+
+import java.util.List;
+public record ErrorResponse(int status, String message, List<String> errors) {
+
+}
