+++ src/main/java/easy/mind/com/api/exception/NullEntityException.java
@@ -1,0 +1,11 @@
+package easy.mind.com.api.exception;
+
+public class NullEntityException extends RuntimeException {
+
+    public NullEntityException() {
+    }
+
+    public NullEntityException(String message) {
+        super(message);
+    }
+}
