+++ src/test/java/easy/mind/com/api/ApiApplicationTests.java
@@ -1,0 +1,13 @@
+package easy.mind.com.api;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
